requirements.txt
flask==2.1.1
flask-sqlalchemy==2.5.1
flask-migrate==3.1.0
psycopg2-binary==2.9.3
SQLAlchemy==1.4.45
Werkzeug==2.0.3
flake8==4.0.1
black==22.3.0
flask-inputs==0.3
jsonschema==4.19.0
python-dotenv==0.10.3
requests==2.23.0
requests-mock==1.10.0
matplotlib==3.7.1
numpy==1.24.0
coverage==7.4.4